MovieRental/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
astroid==3.3.6
click==8.1.7
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
isort==5.13.2
jose==1.0.0
mccabe==0.7.0
passlib==1.7.4
platformdirs==4.3.6
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pylint==3.3.2
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.34.0
